# Add BFloat16 support and optimization for mish, hardtanh backward, and silu on CPU